requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
setuptools>=68.0.0
wheel
channels==4.0.0
channels_redis==4.1.0
daphne==4.0.0
websockets==11.0.3